pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vpn_admin"
version = "0.1.0"
description = "vpn admin on web"
readme = "README.md"
authors = [{name = "PandaRyshan", email = "ryshan@pandas.run"}]
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
dependencies = [
    "Jinja2==3.1.2",
    "Werkzeug==3.0.1",
    "Flask==3.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-WTF==1.2.1",
    "Flask-Admin==1.6.1",
    "Flask-Security-Too==5.3.2",
    "Flask-Caching==2.1.0",
    "Flask-Mail2==1.0.2",
    "email-validator==2.0.0.post2",
    "psycopg2==2.9.9",
    "confluent-kafka==2.3.0",
    "celery==5.3.4",
    "redis==5.0.1",
    "kombu==5.3.2",
    "ansible==8.5.0",
    "WTForms==3.1.1",
    "setuptools==68.2.2",
    "wheel==0.41.3",
    "itsdangerous==2.1.2",
    "python-dotenv==1.0.0"
]

[project.optional-dependencies]
dev = [
    "ruff==0.1.4",
    "ipython==8.17.2"
]
test = [
    "pytest==7.4.3",
    "coverage==7.3.2"
]
prod = ["gunicorn==20.1.0"]
docs = ["sphinx"]

[tool.pytest]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
ignore = ["E115"]
